Introduce soil status threshold
Each time a soil status is changed, the soil status progress is saved on
the flash. If there are too many status changes it can damage the flash.
Soil status threshold allows to wait for some time for a soil status to
stabilize, thus the number of soil status changes will be reduced, and
the corresponding flash writes.